[수정] BottomTableView를 ContainerView로 분리
