+++ OneDay/OneDay.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:OneDay/Model/Device+CoreDataProperties.swift">
-   </FileRef>
-   <FileRef
       location = "group:OneDay.xcodeproj">
    </FileRef>
    <FileRef
