+++ OneDay/OneDay/Controller/SendNotificationProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  SendDataProtocol.swift
+//  OneDay
+//
+//  Created by Wongeun Song on 2019. 2. 15..
+//  Copyright © 2019년 teamA2. All rights reserved.
+//
+
+import Foundation
+
+protocol SendNotificationDelegate {
+    func sendNotification()
+}
